% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact.R
\name{zu_is_secret_header}
\alias{zu_is_secret_header}
\title{Which header names carry a secret value?}
\usage{
zu_is_secret_header(name)
}
\arguments{
\item{name}{A character vector of header names.}
}
\value{
A logical vector.
}
\description{
Which header names carry a secret value?
}
