% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact.R
\name{zu_is_secret_param}
\alias{zu_is_secret_param}
\title{Which query parameter names carry a secret value?}
\usage{
zu_is_secret_param(name)
}
\arguments{
\item{name}{A character vector of parameter names.}
}
\value{
A logical vector.
}
\description{
Which query parameter names carry a secret value?
}
