% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{zu_tls_backend}
\alias{zu_tls_backend}
\title{Which TLS backend was this build linked against?}
\usage{
zu_tls_backend()
}
\value{
A short backend identifier, e.g. \code{"openssl"}, \code{"schannel"},
\code{"sectransport"}.
}
\description{
Which TLS backend was this build linked against?
}
\examples{
zu_tls_backend()
}
